Revert "Extra scraper"
