+++ routes/grades.js
@@ -1,18 +1,36 @@
+var request = require('request');
 var cheerio = require('cheerio');
+var cookieJar = request.jar();
 
-var grades = {
-  url: 'http://micampus.mxl.cetys.mx/portal/auth/portal/default/Academico/Consultar+boleta',
-  parse: function(body){
+function setCookiesOnJar(cookiesString) {
+  var url = "http://micampus.mxl.cetys.mx";
 
-    var i = 3; // first row of table to start 
-    var j = 8; // first column with grades
+  var cookies = cookiesString.split(" ");
 
-    function Course(name) {
-     this.name = name;
-     this.grades = [];
-    }
-    
+  cookies.forEach(function(cookie){
+    cookieJar.setCookie(cookie, url);
+  });
+}
+
+module.exports = function(req, res){
+
+  setCookiesOnJar(res.locals.cookies);
+
+  var getOptions= {
+    url: 'http://micampus.mxl.cetys.mx/portal/auth/portal/default/Academico/Consultar+boleta', 
+    jar: cookieJar
+  };
+  var i = 3; // first row of table to start 
+  var j = 8; // first column with grades
+
+  function Course(name) {
+    this.name = name;
+    this.grades = [];
+  }
+ 
+  request.get(getOptions, function(err, response, body){
     var $ = cheerio.load(body);
+
     var rows = $('table.alumnos-tabla').children();
     var jsonResponse = {courses: []};
 
@@ -26,8 +44,8 @@ var grades = {
       }
       jsonResponse.courses.push(course);
     }
-    return jsonResponse;
-  }
-}
+    console.log(JSON.stringify(jsonResponse));
+    res.send(JSON.stringify(jsonResponse));
 
-module.exports = grades; 
+  });
+}
