+++ routes/schedule.js
@@ -1,28 +1,43 @@
+var request = require('request');
 var cheerio = require('cheerio');
+var cookieJar = request.jar();
 
-var schedule = {
-  url: 'http://micampus.mxl.cetys.mx/portal/auth/portal/default/Academico/Horario',
-  parse: function(body){
-    var i = 0;
-    var j = 1;
+function setCookiesOnJar(cookiesString) {
+  var url = "http://micampus.mxl.cetys.mx";
 
-    function Course(coursecode, name, teacher, groupcode, coursetype) {
-      this.coursecode = coursecode;
-      this.name = name;
-      this.teacher = teacher;
-      this.groupcode = groupcode;
-      this.coursetype = coursetype;
-      this.sessions = [];
-    }
+  var cookies = cookiesString.split(" ");
 
-    function Session(day, houri, hourf, room) {
-      this.day = day;
-      this.houri = houri;
-      this.hourf = hourf;
-      this.room = room;
-    }
+  cookies.forEach(function(cookie){
+    cookieJar.setCookie(cookie, url);
+  });
+}
+module.exports = function(req, res) {
 
-  
+  setCookiesOnJar(res.locals.cookies);
+  var getOptions = {
+    url: 'http://micampus.mxl.cetys.mx/portal/auth/portal/default/Academico/Horario',
+    jar: cookieJar
+  };
+  var i = 0;
+  var j = 1;
+
+  function Course(coursecode, name, teacher, groupcode, coursetype) {
+    this.coursecode = coursecode;
+    this.name = name;
+    this.teacher = teacher;
+    this.groupcode = groupcode;
+    this.coursetype = coursetype;
+    this.sessions = [];
+  }
+
+  function Session(day, houri, hourf, room) {
+    this.day = day;
+    this.houri = houri;
+    this.hourf = hourf;
+    this.room = room;
+  }
+
+  request.get(getOptions, function(err, response, body) {
     var $ = cheerio.load(body);
     var rows = $('table').eq(4).children('tr:has(td[colspan!="7"])');
     if(rows.is('tbody'))
@@ -85,8 +100,6 @@ var schedule = {
       jsonResponse.courses.push(course);
     }
     console.log(JSON.stringify(jsonResponse));
-    return jsonResponse;
-  }
-};
-module.exports = schedule;
-  
+    res.send(JSON.stringify(jsonResponse));
+  });
+}
